+++ Playground/Test.Project/Class1.cs
@@ -1,0 +1,6 @@
+﻿namespace Test.Project;
+
+public class Class1
+{
+
+}